+++ src/app/(tabs)/_layout.tsx
@@ -56,6 +56,7 @@ export default function TabsLayout() {
 
   return (
     <Tabs
+      // initialRouteName="home"
       screenOptions={{
         ...defaultHeaderOptions,
         tabBarStyle: {
