+++ src/app/index.tsx
@@ -1,23 +1,5 @@
-import { Redirect, useRootNavigationState } from 'expo-router';
-import { useAuth } from '../context/auth';
+import { View } from 'react-native';
 
 export default function RootPage() {
-  const { status } = useAuth();
-
-  // Workaround to avoid navigating before the navigator is ready (seen from https://github.com/expo/router/issues/740#issuecomment-1626015285)
-  const rootNavigationState = useRootNavigationState();
-
-  // Render nothing while we are checking auth
-  if (
-    status === 'undetermined' ||
-    status === 'determining' ||
-    !rootNavigationState?.key
-  )
-    return null;
-
-  if (status === 'authenticated') {
-    return <Redirect href="/home" />;
-  }
-
-  return <Redirect href="(auth)/landing" />;
+  return <View style={{ backgroundColor: 'red', flex: 1 }} />;
 }
